Fix for Gump Studio POL Export plug-in
A big thank you to Turley for giving me help to fix the POL plug-in for
Gump Studio. It now exports the text colour correctly. Also in this
commit I am removing extraneous files from the PDG directory.,
specifically the backup folder and files generated by Visual Studio when
it upgraded the solution to VS 2015.